Add historical season import script and fix points to Float
seed_historical.py imports the full 2025-2026 season from the Week 22
scoring spreadsheet: creates the season, teams, bowler/roster entries,
individual game scores (MatchupEntry) for all 21 completed weeks, and
team points (TeamPoints) from the 'team scoring' sheet. Week 22 position
night is created but left unentered since it hasn't been bowled yet.

Also changes TeamPoints.points_earned from Integer to Float so half-point
tied games (e.g. 1.5 pts) are stored accurately — the spreadsheet shows
these in weeks where game scores are exactly tied.

+++ models.py
@@ -202,7 +202,7 @@ class TeamPoints(db.Model):
     week_num = db.Column(db.Integer, nullable=False)
     matchup_num = db.Column(db.Integer, nullable=False)
     team_id = db.Column(db.Integer, db.ForeignKey('teams.id'), nullable=False)
-    points_earned = db.Column(db.Integer, default=0)
+    points_earned = db.Column(db.Float, default=0)  # Float to handle 0.5-pt ties
     is_forfeit = db.Column(db.Boolean, default=False)
 
     team = db.relationship('Team')
